Time: 7 ms (76.19%), Space: 26.1 MB (49.3%) - LeetHub

+++ 1408-find-the-smallest-divisor-given-a-threshold/1408-find-the-smallest-divisor-given-a-threshold.cpp
@@ -4,7 +4,7 @@ public:
 bool helper(vector<int>& nums, int t, int mid){
     long long total=0;
     for(int i=0; i<nums.size(); i++){
-        total+=((long long)nums[i]+(long long)mid-(long long)1)/(long long)mid;
+        total+=(nums[i]+mid-1LL)/mid;
         if(total>t) return 0;
     }
     return 1;
